refactor: file struct change
